All proxy logic was moved to carina-proxy module; carina-webdriver was removed from dependencies list of carina-api
