+++ carina-api/src/main/java/com/qaprosoft/carina/core/foundation/api/AbstractApiMethod.java
@@ -41,13 +41,13 @@ import com.jayway.restassured.filter.log.ResponseLoggingFilter;
 import com.jayway.restassured.http.ContentType;
 import com.jayway.restassured.response.Response;
 import com.jayway.restassured.specification.RequestSpecification;
+import com.qaprosoft.carina.core.foundation.api.http.HttpClient;
+import com.qaprosoft.carina.core.foundation.api.http.HttpMethodType;
+import com.qaprosoft.carina.core.foundation.api.http.HttpResponseStatusType;
 import com.qaprosoft.carina.core.foundation.api.log.LoggingOutputStream;
 import com.qaprosoft.carina.core.foundation.api.ssl.NullHostnameVerifier;
 import com.qaprosoft.carina.core.foundation.api.ssl.NullX509TrustManager;
 import com.qaprosoft.carina.core.foundation.api.ssl.SSLContextBuilder;
-import com.qaprosoft.carina.core.foundation.http.HttpClient;
-import com.qaprosoft.carina.core.foundation.http.HttpMethodType;
-import com.qaprosoft.carina.core.foundation.http.HttpResponseStatusType;
 import com.qaprosoft.carina.core.foundation.utils.Configuration;
 import com.qaprosoft.carina.core.foundation.utils.Configuration.Parameter;
 import com.qaprosoft.carina.core.foundation.utils.R;
