+++ carina-proxy/pom.xml
@@ -29,6 +29,11 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>com.qaprosoft</groupId>
+			<artifactId>carina-utils</artifactId>
+		</dependency>
+		
+		<dependency>
 			<groupId>log4j</groupId>
 			<artifactId>log4j</artifactId>
 		</dependency>
